*extend can timestamp to 64bit ( actually we use only 48bit ) *set sequential can message order

+++ Src/xcan_can.h
@@ -16,7 +16,7 @@ typedef struct
   uint8_t   dlc;
   uint8_t   flags;
   uint8_t   dummy;
-  uint32_t timestamp;
+  uint64_t  timestamp;
 }
 can_message_t;
